refactor(user-dashboard): modularize dashboard architecture
